🧹 Fix: Code Health - Resolve code duplication in processor.py
